+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.1;
-const unsigned char AutoSDKVersionString[] = "1.1.2";
+double AutoSDKVersionNumber = 1.2.0;
+const unsigned char AutoSDKVersionString[] = "1.2.0";